Move conf structs to header
